defrag: make frag_end and tracker id conversions explicit
-fsanitize=implicit-conversion flags frag_offset + data_len narrowing into the uint16_t frag_end. For IPv4 compute frag_end after the existing too-large check so the sum always fits. For IPv6 there is no
equivalent check: keep the historic wrap-around explicit and leave a TODO, since RFC 8200 4.5 requires dropping fragments whose reassembled
size would exceed 65535 bytes (needs its own ticket and SV test).

Also drop the vestigial int32 casts when assigning the uint32_t tracker id.

Ticket: 8232

+++ src/defrag-hash.c
@@ -130,11 +130,11 @@ static void DefragTrackerInit(DefragTracker *dt, Packet *p)
 
     if (PacketIsIPv4(p)) {
         const IPV4Hdr *ip4h = PacketGetIPv4(p);
-        dt->id = (int32_t)IPV4_GET_RAW_IPID(ip4h);
+        dt->id = IPV4_GET_RAW_IPID(ip4h);
         dt->af = AF_INET;
     } else {
         DEBUG_VALIDATE_BUG_ON(!PacketIsIPv6(p));
-        dt->id = (int32_t)IPV6_EXTHDR_GET_FH_ID(p);
+        dt->id = IPV6_EXTHDR_GET_FH_ID(p);
         dt->af = AF_INET6;
     }
     dt->proto = PacketGetIPProto(p);
